chore: remove event calendar
